requirements.txt
openai-whisper==20231117
torch==2.1.2
pyannote.audio==2.1.1
yt-dlp==2024.3.4
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.5.0
easyocr==1.6.2
opencv-python==4.8.1.78
